Bug 1327714: MySQL cartridge removes logs on start
The "start" function of start script of the MySQL cartridge removes the
mysql.log which is supposed to save some disk space. However, since the
log file is removed, it's very difficult, if not impossible, for users
to debug any issues with MySQL database.

Since the log file removal is unnecessary, the code that is responsible
for that is now removed. As a result, the mysql.log file will remain
after the "start" function in control script is called. Also, replace
an error-log dump with a static message to instruct users to check
mysql.log for errors.

Bug: 1327714
Link: <https://bugzilla.redhat.com/show_bug.cgi?id=1327714>

Signed-off-by: Vu Dinh <vdinh@redhat.com>