Openshift doesn't accept app aliases containing an underscore
Currently, application aliases with underscore(s) are not allowed due
to regex restriction during alias verification
(controller/app/models/application.rb validate_alias(fqdn)).

Having underscore(s) in URL is generally not recommended but it's not
technically wrong to have. So, the regex for validate_alias method
is modified to accept underscore(s).

Bug: 1329915
Link: https://bugzilla.redhat.com/show_bug.cgi?id=1329915

Signed-off-by: Vu Dinh <vdinh@redhat.com>