Fix documentation links to point to origin-server
Documentation links previously pointing to openshift.github.io are
broken and must be changed to point to the corresponding adoc file in
the github origin-server repository.

Any links to documentation not provided in the origin-server repository
(such as the api guide) have been changed to point to the publicly
available enterprise documentation.