routing-daemon: Resynchronize with load-balancer
Resynchronize the routing daemon with the load balancer when the daemon
attempts to perform on operation on an unrecognized pool or pool member.

The intention of this commit is to improve the daemon's reliability when
something other than the current instance of the daemon changes the load
balancer's state, for example the system administrator or clustered routing
daemons.

AsyncLoadBalancerController::Pool, BatchedLoadBalancerController::Pool,
LoadBalancerController::Pool, SimpleLoadBalancerController::Pool: Add
synch method.

AsyncLoadBalancerController, BatchedLoadBalancerController,
LoadBalancerController, SimpleLoadBalancerController: Add synch_pools
method.

RoutingDaemon: Add with_pool method, which checks whether a pool exists,
resynchronizes the daemon with the load balancer if it is not found (using
the new synch_pools controller method), optionally creates the pool if it
is still not found, and invokes the supplied block if the pool is found.

RoutingDaemon#create_application: Take pool_name as a parameter rather than
generating the value itself since callers will already have generated the
value themselves.

RoutingDaemon#delete_application, RoutingDaemon#add_endpoint,
RoutingDaemon#remove_endpoint, RoutingDaemon#add_alias,
RoutingDaemon#remove_alias, RoutingDaemon#add_ssl,
RoutingDaemon#remove_ssl: Use the daemon's new with_pool method in order to
check whether the pool exists, resynchronize the daemon with the load
balancer if an operation is attempted on an unrecognized pool, and (in the
case of add methods) create the pool if it does not exist.

RoutingDaemon#remove_endpoint: Use the controller's new synch method for
pools to resynchronize the daemon with the load balancer if the operation
is attempted on an unknown pool member.