Revert dfac396 and enable routing notifications during moves
The routing daemon was previously not receiving updated endpoint
notifications during gear moves. A previous commit, dfac396, enabled
endpoints to be published but the endpoints published were the same
endpoints as before the move.

This change reverts dfac396 and properly sends endpoint creation and
deletion events to the routing daemon during a gear move.