routing-daemon: F5: Delete client-ssl profile
F5IControlRestLoadBalancerModel#remove_ssl: Try to delete the client-ssl
profile even if deleting it from the virtual server fails; try to delete
the key even if deleting the client-ssl fails; and try to delete the
certificate even if deleting the key fails.

LBModelExceptionCollector: New helper class to rescue, collect, and combine
LBModelException exceptions.

F5 BIG-IP will not let us delete a resource that is in use (e.g.,
a client-ssl profile that is still associated with a virtual server), but
this change will enable the routing-daemon to finish cleaning up
a client-ssl profile or associated key or certificate that is already
deleted from the virtual server.