Fix input/output deadlock condition
If the rate of input and output are very high, the message queue size
can trend towards zero. In this case, it's possible that the Input will
decide to evict from the queue, but by the time the eviction occurs the
output has drained the queue to zero, causing the eviction dequeue to
block forever. In this case, Input is now blocked trying to evict, and
Output is blocked waiting for Input.

Fix the deadlock by distinguishing between eviction of old messages and
dropping of new messages, and making those operations non-blocking.

Add a test which verifies the fix under these conditions.